refactor(ui): Split WecsDetailsPanel into modular tab components for better maintainability (#1555)
* Refractor of WecsDetailsPanel

* Fixes Minor error

* Fixes

* Fixes YAML issue

* Fixes lint

* fixes missing functionability

* Fixes Conflicts

* Fixes Lint
